backend/requirements.txt
asgiref==3.9.1
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
dj-database-url==2.1.0
Django==5.2.5
django-cors-headers==4.4.0
django-extensions==4.1
djangorestframework==3.15.2
idna==3.10
iniconfig==2.1.0
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
psycopg==3.2.9
psycopg-binary==3.2.9
pydotplus==2.0.2
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.2
pytest-django==4.11.1
python-dotenv==1.1.1
requests==2.32.5
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
pybreaker==1.1.0
lxml==5.2.2; python_version < "3.13"
lxml>=5.4,<6; python_version >= "3.13"
pydantic>=2.0


weasyprint 
